move successrate calculation and feedback message to utility file
