+++ src/main/js/components/guessplaypage.js
@@ -2,6 +2,7 @@ import React from 'react';
 import {ButtonToolbar, Grid, Row, Col} from 'react-bootstrap';
 import 'whatwg-fetch';
 import AnswerButton from './answerbutton';
+import Utility from '../util/utility';
 
 export default class GuessPlayPage extends React.Component {
     constructor(props) {
@@ -97,35 +98,6 @@ export default class GuessPlayPage extends React.Component {
             setTimeout(() => this.props.switchPage('EndScreenPage', '', 'GuessPlayPage', this.state.results), 1000);
         }
     }
-    getSuccessRate(){
-        let successRate = 0;
-        let successRateMessage = '';
-        const emojiFeedback = {
-            veryBad : String.fromCodePoint(0x1F61E),
-            bad : String.fromCodePoint(0x1F615),
-            average : String.fromCodePoint(0x1F610),
-            good : String.fromCodePoint(0x1F642),
-            veryGood : String.fromCodePoint(0x1F600)
-        };
-        if (Number(sessionStorage.totalAttempts) > 0) {
-            successRate = Number(sessionStorage.correctAttempts)
-                / Number(sessionStorage.totalAttempts) * 100;
-            successRateMessage = `${successRate.toFixed(0)} %`;
-            if(successRate >= 80){
-                return `, ${successRateMessage} ${emojiFeedback['veryGood']}`;
-            } else if(successRate < 80 && successRate >= 60){
-                return `, ${successRateMessage} ${emojiFeedback['good']}`;
-            } else if(successRate < 60 && successRate >= 40){
-                return `, ${successRateMessage} ${emojiFeedback['average']}`;
-            } else if(successRate < 40 && successRate >= 20){
-                return `, ${successRateMessage} ${emojiFeedback['bad']}`;
-            } else if(successRate < 20){
-                return `, ${successRateMessage} ${emojiFeedback['veryBad']}`;
-            }
-        } else {
-              return successRateMessage;
-        }
-    }
     onKeys(event){
         let keyDown = event.key;
         if(!this.state.buttonDisabled){
@@ -186,7 +158,7 @@ export default class GuessPlayPage extends React.Component {
                             Fråga: {(Number(sessionStorage.currentQuestionIndex) + 1) + ' / '
                             + this.state.lessonLength}
                             <br/>
-                            {sessionStorage.correctAttempts + ' rätt' + this.getSuccessRate()}
+                            {sessionStorage.correctAttempts + ' rätt' + Utility.getSuccessRate()}
                         </div>
                     </Row>
                 </Grid>
