+++ src/main/js/components/translationplaypage.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import {Button, Grid, Row} from 'react-bootstrap';
 import 'whatwg-fetch';
+import Utility from '../util/utility';
 
 export default class TranslationPlayPage extends React.Component {
     constructor(props) {
@@ -70,35 +71,6 @@ export default class TranslationPlayPage extends React.Component {
             setTimeout(() => this.props.switchPage('EndScreenPage', '', 'TranslationPlayPage', this.state.results), 2000);
         }
     }
-    getSuccessRate(){
-        let successRate = 0;
-        let successRateMessage = '';
-        const emojiFeedback = {
-            veryBad : String.fromCodePoint(0x1F61E),
-            bad : String.fromCodePoint(0x1F615),
-            average : String.fromCodePoint(0x1F610),
-            good : String.fromCodePoint(0x1F642),
-            veryGood : String.fromCodePoint(0x1F600)
-        };
-        if (Number(sessionStorage.totalAttempts) > 0) {
-            successRate = Number(sessionStorage.correctAttempts)
-                / Number(sessionStorage.totalAttempts) * 100;
-            successRateMessage = `${successRate.toFixed(0)} %`;
-            if(successRate >= 80){
-                return `, ${successRateMessage} ${emojiFeedback['veryGood']}`;
-            } else if(successRate < 80 && successRate >= 60){
-                return `, ${successRateMessage} ${emojiFeedback['good']}`;
-            } else if(successRate < 60 && successRate >= 40){
-                return `, ${successRateMessage} ${emojiFeedback['average']}`;
-            } else if(successRate < 40 && successRate >= 20){
-                return `, ${successRateMessage} ${emojiFeedback['bad']}`;
-            } else if(successRate < 20){
-                return `, ${successRateMessage} ${emojiFeedback['veryBad']}`;
-            }
-        } else {
-              return successRateMessage;
-        }
-    }
     render() {
         return (
             <div>
@@ -119,7 +91,7 @@ export default class TranslationPlayPage extends React.Component {
                             Fråga: {(Number(sessionStorage.currentQuestionIndex) + 1) + ' / '
                             + this.state.lessonLength}
                             <br/>
-                            {sessionStorage.correctAttempts + ' rätt' + this.getSuccessRate()}
+                            {sessionStorage.correctAttempts + ' rätt' + Utility.getSuccessRate()}
                         </div>
                     </Row>
                     <Row><h3>{this.state.output}</h3></Row>
